auth added code

+++ app/Http/Controllers/Admin/UsersController.php
@@ -8,7 +8,8 @@ use Illuminate\Http\Request;
 use App\Models\Admin\Permission;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
-
+use Gate;
+use Symfony\Component\HttpFoundation\Response;
 class UsersController extends Controller
 {
     /**
@@ -16,7 +17,9 @@ class UsersController extends Controller
      */
     public function index()
     {
-         // users data with order by id desc
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+         
+        // users data with order by id desc
         $users = User::orderBy('id', 'desc')->with('roles')->get();
         return response()->view('admin.users.index', compact('users'));
     }
@@ -26,6 +29,8 @@ class UsersController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        
         $roles = Role::all()->pluck('title', 'id');
         return response()->view('admin.users.create', compact('roles'));
     }
@@ -50,7 +55,9 @@ class UsersController extends Controller
      */
     public function show(string $id)
     {
-    $user_detail = User::with(['roles', 'roles.permissions'])->findOrFail($id);
+        abort_if(Gate::denies('user_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+    
+        $user_detail = User::with(['roles', 'roles.permissions'])->findOrFail($id);
     $roles = Role::all();
     $permissions = Permission::all();
     return view('admin.users.show', compact('user_detail', 'roles', 'permissions'));
@@ -61,7 +68,9 @@ class UsersController extends Controller
      */
     public function edit(string $id)
     {
-         $user_edit = User::find($id);
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+         
+        $user_edit = User::find($id);
         $roles = Role::all()->pluck('title', 'id');
         return response()->view('admin.users.edit', compact('user_edit', 'roles'));
     }
@@ -82,7 +91,9 @@ class UsersController extends Controller
      */
     public function destroy(string $id)
     {
-         $user = User::find($id);
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        
+        $user = User::find($id);
         $user->delete();
         return redirect()->route('admin.users.index')->with('success', 'User deleted successfully');
     }
